Updated SKUser
Added all new properties, note some of the seemingly obvious/standard fields are not in the default field (e.g. answerCount)

Badge count is in a dictionary, not worth having a Badge Counts object. The code for the badge counts is general and can be moved into the super method.